feat: #903 첫 사용자 온보딩 상태 관리 기능 추가 및 관련 컴포넌트 수정
