+++ apps/web/src/app/desktop/component/home/Onboarding/index.tsx
@@ -1,7 +1,4 @@
 import { css } from "@emotion/react";
-import { useState, useEffect } from "react";
-import Cookies from "js-cookie";
-import { COOKIE_KEYS } from "@/config/storage-keys";
 import { Icon } from "@/component/common/Icon";
 import { Typography } from "@/component/common/typography";
 import { DESIGN_TOKEN_COLOR } from "@/style/designTokens";
@@ -26,63 +23,7 @@ const ONBOARDING_STEPS = [
   },
 ];
 
-const ONBOARDING_STORAGE_KEY = "layer_home_onboarding_closed";
-
-/**
- * 온보딩 닫기 기록 불러오기
- *
- * @returns Record<string, boolean>
- */
-const getOnboardingClosedRecord = (): Record<string, boolean> => {
-  const stored = localStorage.getItem(ONBOARDING_STORAGE_KEY);
-  return stored ? JSON.parse(stored) : {};
-};
-
-/**
- * 특정 사용자가 온보딩을 닫았는지 확인
- *
- * @param memberId
- * @returns boolean
- */
-const hasUserClosedOnboarding = (memberId: string): boolean => {
-  const record = getOnboardingClosedRecord();
-  return record[memberId] === true;
-};
-
-/**
- * 특정 사용자의 온보딩 닫기 기록 저장
- *
- * @param memberId
- */
-const saveOnboardingClosed = (memberId: string): void => {
-  const record = getOnboardingClosedRecord();
-  record[memberId] = true;
-  localStorage.setItem(ONBOARDING_STORAGE_KEY, JSON.stringify(record));
-};
-
-export default function Onboarding() {
-  const [isVisible, setIsVisible] = useState(false);
-  const memberId = Cookies.get(COOKIE_KEYS.memberId);
-
-  const handleClose = () => {
-    if (!memberId) return;
-
-    saveOnboardingClosed(memberId);
-    setIsVisible(false);
-  };
-
-  useEffect(() => {
-    if (!memberId) {
-      setIsVisible(false);
-      return;
-    }
-
-    const shouldShowOnboarding = !hasUserClosedOnboarding(memberId);
-    setIsVisible(shouldShowOnboarding);
-  }, [memberId]);
-
-  if (!isVisible) return null;
-
+export default function Onboarding({ onClose }: { onClose: () => void }) {
   return (
     <article
       css={css`
@@ -117,7 +58,7 @@ export default function Onboarding() {
         </Typography>
 
         <button
-          onClick={handleClose}
+          onClick={onClose}
           css={css`
             border: none;
             display: flex;
