+++ apps/web/src/app/desktop/home/HomePage.tsx
@@ -5,8 +5,11 @@ import InProgressRetrospectsWrapper from "@/app/desktop/component/home/InProgres
 import ActionItemsWrapper from "@/app/desktop/component/home/ActionItemsWrapper";
 import AnalyticsWrapper from "../component/home/AnalyticsWrapper";
 import Onboarding from "../component/home/Onboarding";
+import { useFirstTimeUser } from "../component/home/Onboarding/useFirstTimeUser";
 
 export function HomePage() {
+  const { isFirstTimeUser, dismiss: dismissFirstTimeUser } = useFirstTimeUser();
+
   return (
     <section
       css={css`
@@ -21,10 +24,10 @@ export function HomePage() {
       <HomePageHeader />
 
       {/* ---------- 온보딩 ---------- */}
-      <Onboarding />
+      {isFirstTimeUser && <Onboarding onClose={dismissFirstTimeUser} />}
 
-      {/* ---------- 작성중인 회고 ---------- */}
-      <InProgressRetrospectsWrapper />
+      {/* ---------- 작성중인 회고 (첫 사용자에게는 숨김) ---------- */}
+      {!isFirstTimeUser && <InProgressRetrospectsWrapper />}
 
       {/* ---------- 실행목표 ---------- */}
       <ActionItemsWrapper />
